add decimal fields to models

+++ api/models.py
@@ -26,7 +26,7 @@ class CocktailIngredient(models.Model):
     cocktail = models.ForeignKey(Cocktail, on_delete=models.PROTECT)
     ingredient = models.ForeignKey(Ingredient, on_delete=models.PROTECT)
     sort_order = models.PositiveIntegerField()
-    amount = models.PositiveIntegerField()
+    amount = models.DecimalField(max_digits=10, decimal_places=2)
     unit = models.CharField(max_length=120)
 
     def __str__(self):
@@ -67,7 +67,7 @@ class UserProduct(models.Model):
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     product = models.ForeignKey(Product, null=True, blank=True, on_delete=models.CASCADE)
     quantity = models.PositiveIntegerField(default=1)
-    amount_available = models.PositiveIntegerField()
+    amount_available = models.DecimalField(max_digits=25, decimal_places=2)
 
     def __str__(self):
         return f"{self.product.name} in {self.user.username}"
